Implemented product service and controller

+++ src/main/java/com/scaler/amit/project_ecomservice/repositories/OrderRepository.java
@@ -1,9 +1,9 @@
 package com.scaler.amit.project_ecomservice.repositories;
 
-import org.springframework.core.annotation.Order;
+import com.scaler.amit.project_ecomservice.models.order.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrderRepository extends JpaRepository<Order, Integer> {
+public interface OrderRepository extends JpaRepository<Order, Long> {
 }
